fix: fix import and improve types
