+++ packages/sdk/src/codegen/templates/typescript/jsonschema.ts
@@ -1,6 +1,6 @@
 import { doNotEditHeader, Template, TemplateOutputFile } from '../../index';
 import { ResolvedWunderGraphConfig } from '../../../configure';
-import { formatTypeScript, loadFile } from './index';
+import { formatTypeScript } from './index';
 import Handlebars from 'handlebars';
 import { template } from './jsonschema.template';
