+++ packages/sdk/src/codegen/templates/typescript/server.ts
@@ -1,7 +1,7 @@
 import { doNotEditHeader, Template, TemplateOutputFile } from '../../index';
 import { ResolvedWunderGraphConfig } from '../../../configure';
 import Handlebars from 'handlebars';
-import { formatTypeScript, TypeScriptInputModels, TypeScriptResponseModels } from './index';
+import { formatTypeScript } from './index';
 import { template } from './server.template';
 import { WunderGraphHooksPlugin } from './hooks';
 import { WunderGraphInternalApiClient } from './internal.client';
