+++ packages/vite-testapp/.wundergraph/wundergraph.config.ts
@@ -134,7 +134,7 @@ configureWunderGraphApplication({
 	],
 	cors: {
 		...cors.allowAll,
-		allowedOrigins: ['http://localhost:5173', 'http://127.0.0.1:5173'],
+		allowedOrigins: ['http://127.0.0.1:5173'],
 	},
 	authentication: {
 		cookieBased: {
