+++ packages/vite-testapp/package.json
@@ -8,7 +8,7 @@
     "wundergraph": "wunderctl up --debug",
     "generate": "wunderctl generate",
     "browser": "wait-on \"http-get://localhost:5173\" && wait-on \"http-get://localhost:9991\" && open-cli http://localhost:5173",
-    "dev:vite": "vite",
+    "dev:vite": "vite --force",
     "check": "tsc --noEmit",
     "build": "pnpm generate && tsc && vite build",
     "preview": "vite preview"
